Moved function to the correct file
